code docs updated
