+++ README.md
@@ -1,11 +1,17 @@
+# React Native Shapes
 [![Build Status](https://travis-ci.org/pavinthan/react-native-shapes.svg?branch=master)](https://travis-ci.org/pavinthan/react-native-shapes)
 
-This project was bootstrapped with [Create React Native Root](https://github.com/react-community/create-react-native-app).
+This mobile application will generate different shapes by using users natural language.
 
+This project was bootstrapped with [Create React Native Root](https://github.com/react-community/create-react-native-app).
 Below you'll find information about performing common tasks.
 
 ## Table of Contents
 
+- [Getting Started](#getting-started)
+  - [Run the application](#run-the-application)
+  - [Hint & Example](#hint-example)
+  - [Supporting shapes](#supporting-shapes)
 - [Available Scripts](#available-scripts)
   - [npm start](#npm-start)
   - [npm test](#npm-test)
@@ -13,6 +19,60 @@ Below you'll find information about performing common tasks.
   - [npm run android](#npm-run-android)
   - [npm run eject](#npm-run-eject)
 - [Writing and Running Tests](#writing-and-running-tests)
+
+## Getting Started
+This code will run using Expo without rebuilding on  Android or IOS
+
+### Run the application:
+Open project in react native compatible IDE
+Some basic commands to use in Command Prompt or Terminal
+```
+yarn start
+```    
+Starts the development server so you can open your app in the Expo app on your phone.
+
+```
+  yarn run ios
+```
+Starts the development server and loads your app in an iOS simulator. (Mac only, requires Xcode)
+
+```
+  yarn run android
+```    
+Starts the development server and loads your app on a connected Android device or emulator. (Requires Android build tools)
+
+```
+yarn test
+```
+Starts the test runner
+
+## Hint & Example
+### Hint to get a best result
+```
+Draw a(n) <shape> with a(n) <measurement> of <amount> (and a(n) <measurement> of <amount>)
+```
+
+### Examples
+```
+Draw a square with a side length of 200 
+Draw a rectangle with a width of 250 and a height of 400 
+Draw an octagon with a side length of 200 
+Draw an isosceles triangle with a height of 200 and a width of 100
+```
+
+## Supporting shapes
+-	[x] Isosceles Triangle 	
+-	[x]	Square 
+-	[x]	Scalene Triangle 	
+-	[x]	Parallelogram 
+-	[x]	Equilateral Triangle 	
+-	[x]	Pentagon 
+-	[x]	Rectangle
+-	[x]	Hexagon 
+-	[ ]	Heptagon – (currently not available)
+-	[x]	Octagon 
+-	[x]	Circle 
+-	[x]	Oval 
 
 ## Available Scripts
