+++ native-base-theme/variables/platform.js
@@ -5,7 +5,8 @@ const deviceHeight = Dimensions.get('window').height;
 const deviceWidth = Dimensions.get('window').width;
 const platform = Platform.OS;
 const platformStyle = undefined;
-const isIphoneX = platform === 'ios' && deviceHeight === 812 && deviceWidth === 375;
+const isIphoneX =
+  platform === 'ios' && deviceHeight === 812 && deviceWidth === 375;
 
 export default {
   platformStyle,
