+++ src/helpers/string.js
@@ -1,3 +1,9 @@
+/**
+ * Find string between two string string.
+ * @param string
+ * @param start
+ * @param end
+ */
 export const strBetween = (string, start, end) => {
   // create regular expression
   const regex = new RegExp(
